new Change

+++ rest-service/src/main/java/uz/pdp/restservice/controller/AgentMerchantController.java
@@ -3,19 +3,16 @@ package uz.pdp.restservice.controller;
 import admin.receive.AgentMerchantReceiveDto;
 import admin.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
-import uz.pdp.restservice.model.AgentMerchantEntity;
 import uz.pdp.restservice.service.base.BaseService;
 
-import javax.validation.Valid;
 import java.util.Optional;
 
 import static uz.pdp.restservice.service.base.ResponseMessage.ID_EMPTY;
+import static uz.pdp.restservice.service.base.ResponseMessage.SUCCESS;
 
 @RestController
-@RequestMapping("/agentMerchant")
+@RequestMapping("/agent/merchant")
 public class AgentMerchantController {
 
     private final BaseService baseService;
@@ -25,37 +22,37 @@ public class AgentMerchantController {
     }
 
     @PostMapping("/add")
-    public ApiResponse add(@Valid @RequestBody AgentMerchantReceiveDto agentMerchantReceiveDto) {
+    public ApiResponse add(@RequestBody AgentMerchantReceiveDto agentMerchantReceiveDto) {
         return baseService.add(agentMerchantReceiveDto);
     }
 
-    @GetMapping("/get")
-    public ApiResponse getList() {
-        return baseService.getList();
+    @PostMapping("/list/{page}")
+    public ApiResponse getList(@PathVariable int page, @RequestBody AgentMerchantReceiveDto agentMerchantReceiveDto) {
+//        if (baseService.isNotNull(agentMerchantReceiveDto)) {
+//            return baseService.getList(page, agentMerchantReceiveDto);
+//        }
+        return baseService.getList(page, agentMerchantReceiveDto);
     }
 
-    @GetMapping("/get/disabled")
-    public ApiResponse getDisabledList() {
-        return baseService.getDisabledList();
+
+    @GetMapping("/list/all")
+    public ApiResponse getAllList() {
+        return baseService.getAllList();
     }
 
-    @PutMapping("/edit/{id}")
-    public ApiResponse edit(@PathVariable Long id,@Valid @RequestBody AgentMerchantReceiveDto agentMerchantReceiveDto) {
-        return baseService.edit(id, agentMerchantReceiveDto);
+    @GetMapping("/get/{id}")
+    public ApiResponse get(@PathVariable long id) {
+        return new ApiResponse<>(0, SUCCESS, baseService.getById(id));
     }
 
-    @GetMapping("/getById/{id}")
-    public AgentMerchantEntity getById(@PathVariable Long id){
-        return (AgentMerchantEntity) baseService.getById(id);
+    @PostMapping("/edit/{id}")
+    public ApiResponse edit(@PathVariable long id, @RequestBody AgentMerchantReceiveDto agentMerchantReceiveDto) {
+        return baseService.edit(agentMerchantReceiveDto, id);
     }
-
 
     @DeleteMapping("/delete/{id}")
     public ApiResponse delete(@PathVariable Optional<Long> id) {
-        if (id.isPresent())
-            return baseService.delete(id.get());
-
-        else
-            return new ApiResponse<>(1, ID_EMPTY);
+        if (id.isPresent()) return baseService.delete(id.get());
+        else return new ApiResponse<>(1, ID_EMPTY);
     }
 }
