+++ rest-service/src/main/java/uz/pdp/restservice/controller/GatewayMerchantController.java
@@ -1,24 +1,19 @@
 package uz.pdp.restservice.controller;
 
-import admin.receive.AgentMerchantReceiveDto;
 import admin.receive.AgentReceiveDto;
 import admin.receive.GatewayMerchantReceiveDto;
 import admin.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import uz.pdp.restservice.model.AgentMerchantEntity;
-import uz.pdp.restservice.model.GatewayEntity;
-import uz.pdp.restservice.model.GatewayMerchantEntity;
 import uz.pdp.restservice.service.base.BaseService;
 
-import javax.validation.Valid;
 import java.util.Optional;
 
 import static uz.pdp.restservice.service.base.ResponseMessage.ID_EMPTY;
+import static uz.pdp.restservice.service.base.ResponseMessage.SUCCESS;
 
 @RestController
-@RequestMapping("/gatewayMerchant")
+@RequestMapping("/gateway/merchant")
 public class GatewayMerchantController {
 
     private final BaseService baseService;
@@ -28,36 +23,36 @@ public class GatewayMerchantController {
     }
 
     @PostMapping("/add")
-    public ApiResponse add(@Valid @RequestBody GatewayMerchantReceiveDto gatewayMerchantReceiveDto) {
+    public ApiResponse add(@RequestBody GatewayMerchantReceiveDto gatewayMerchantReceiveDto) {
         return baseService.add(gatewayMerchantReceiveDto);
     }
 
-    @GetMapping("/get")
-    public ApiResponse getList() {
-        return baseService.getList();
+    @PostMapping("/list/{page}")
+    public ApiResponse getList(
+            @PathVariable int page,
+            @RequestBody GatewayMerchantReceiveDto gatewayMerchantReceiveDto
+    ) {
+        return baseService.getList(page, gatewayMerchantReceiveDto);
     }
 
-    @GetMapping("/get/disabled")
-    public ApiResponse getDisabledList() {
-        return baseService.getDisabledList();
+    @GetMapping("/list/all")
+    public ApiResponse getAllList() {
+        return baseService.getAllList();
     }
 
-    @PutMapping("/edit/{id}")
-    public ApiResponse edit(@PathVariable Long id,@Valid @RequestBody GatewayMerchantReceiveDto gatewayMerchantReceiveDto) {
-        return baseService.edit(id, gatewayMerchantReceiveDto);
+    @GetMapping("/get/{id}")
+    public ApiResponse get(@PathVariable long id) {
+        return new ApiResponse<>(0, SUCCESS, baseService.getById(id));
     }
 
-    @GetMapping("/getById/{id}")
-    public GatewayMerchantEntity getById(@PathVariable Long id){
-        return (GatewayMerchantEntity) baseService.getById(id);
+    @PostMapping("/edit/{id}")
+    public ApiResponse edit(@PathVariable long id, @RequestBody GatewayMerchantReceiveDto gatewayMerchantReceiveDto) {
+        return baseService.edit(gatewayMerchantReceiveDto, id);
     }
 
     @DeleteMapping("/delete/{id}")
     public ApiResponse delete(@PathVariable Optional<Long> id) {
-        if (id.isPresent())
-            return baseService.delete(id.get());
-
-        else
-            return new ApiResponse<>(1, ID_EMPTY);
+        if (id.isPresent()) return baseService.delete(id.get());
+        else return new ApiResponse<>(1, ID_EMPTY);
     }
 }
