+++ rest-service/src/main/java/uz/pdp/restservice/controller/MerchantController.java
@@ -1,18 +1,15 @@
 package uz.pdp.restservice.controller;
 
-import admin.receive.GatewayMerchantReceiveDto;
 import admin.receive.MerchantReceiveDto;
 import admin.response.ApiResponse;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import uz.pdp.restservice.model.GatewayMerchantEntity;
 import uz.pdp.restservice.service.base.BaseService;
 
-import javax.validation.Valid;
 import java.util.Optional;
 
 import static uz.pdp.restservice.service.base.ResponseMessage.ID_EMPTY;
+import static uz.pdp.restservice.service.base.ResponseMessage.SUCCESS;
 
 @RestController
 @RequestMapping("/merchant")
@@ -25,36 +22,31 @@ public class MerchantController {
     }
 
     @PostMapping("/add")
-    public ApiResponse add(@Valid @RequestBody MerchantReceiveDto merchantReceiveDto) {
+    public ApiResponse add(@RequestBody MerchantReceiveDto merchantReceiveDto) {
         return baseService.add(merchantReceiveDto);
     }
 
-    @GetMapping("/get")
-    public ApiResponse getList() {
-        return baseService.getList();
+    @PostMapping("/list/{page}")
+    public ApiResponse getList(@PathVariable int page, @RequestBody MerchantReceiveDto merchantReceiveDto) {
+        if (baseService.isNotNull(merchantReceiveDto)) {
+            return baseService.getList(page, merchantReceiveDto);
+        }
+        return baseService.getList(page);
     }
 
-    @GetMapping("/get/disabled")
-    public ApiResponse getDisabledList() {
-        return baseService.getDisabledList();
+    @GetMapping("/list/all")
+    public ApiResponse getAllList() {
+        return baseService.getAllList();
     }
 
-    @PutMapping("/edit/{id}")
-    public ApiResponse edit(@PathVariable Long id,@Valid @RequestBody MerchantReceiveDto merchantReceiveDto) {
-        return baseService.edit(id, merchantReceiveDto);
-    }
-
-    @GetMapping("/getById/{id}")
-    public MerchantReceiveDto getById(@PathVariable Long id){
-        return (MerchantReceiveDto) baseService.getById(id);
+    @GetMapping("/get/{id}")
+    public ApiResponse get(@PathVariable Optional<Long> id) {
+        return new ApiResponse<>(0, SUCCESS, baseService.getById(id.get()));
     }
 
     @DeleteMapping("/delete/{id}")
     public ApiResponse delete(@PathVariable Optional<Long> id) {
-        if (id.isPresent())
-            return baseService.delete(id.get());
-
-        else
-            return new ApiResponse<>(1, ID_EMPTY);
+        if (id.isPresent()) return baseService.delete(id.get());
+        else return new ApiResponse<>(1, ID_EMPTY);
     }
 }
